Fix: 개별 포트폴리오 조회시 응답 데이터에 종목 로고 추가 및 리팩토링

+++ src/main/java/com/hrth/ustock/service/PortfolioService.java
@@ -3,7 +3,7 @@ package com.hrth.ustock.service;
 import com.hrth.ustock.dto.holding.HoldingEmbedDto;
 import com.hrth.ustock.dto.holding.HoldingRequestDto;
 import com.hrth.ustock.dto.portfolio.PortfolioListDto;
-import com.hrth.ustock.dto.portfolio.PortfolioRequestDTO;
+import com.hrth.ustock.dto.portfolio.PortfolioRequestDto;
 import com.hrth.ustock.dto.portfolio.PortfolioResponseDto;
 import com.hrth.ustock.dto.portfolio.PortfolioUpdateDto;
 import com.hrth.ustock.entity.User;
@@ -80,9 +80,10 @@ public class PortfolioService {
             HoldingEmbedDto holdingEmbedDto = HoldingEmbedDto.builder()
                     .code(stock.getCode())
                     .name(stock.getName())
+                    .logo(stock.getLogo())
                     .quantity(h.getQuantity())
                     .average(h.getAverage())
-                    .ror((h.getQuantity() * h.getAverage() == 0) ? 0.0 : (double) (ret / (h.getQuantity() * h.getAverage())) * 100)
+                    .ror((h.getQuantity() * h.getAverage() == 0) ? 0.0 : (double) ret / ((long) h.getQuantity() * h.getAverage()) * 100)
                     .build();
             portfolioResponseDto.getStocks().add(holdingEmbedDto);
         });
@@ -96,11 +97,11 @@ public class PortfolioService {
     }
 
     @Transactional
-    public void addPortfolio(PortfolioRequestDTO portfolioRequestDTO, Long userId) {
+    public void addPortfolio(PortfolioRequestDto portfolioRequestDto, Long userId) {
 
         User user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
         Portfolio portfolio = Portfolio.builder()
-                .name(portfolioRequestDTO.getName())
+                .name(portfolioRequestDto.getName())
                 .user(user)
                 .budget(0L)
                 .principal(0L)
@@ -147,8 +148,6 @@ public class PortfolioService {
     public void deleteHolding(Long pfId, String code) {
 
         Portfolio portfolio = portfolioRepository.findById(pfId).orElseThrow(PortfolioNotFoundException::new);
-        List<Holding> holdings = portfolio.getHoldings();
-
         Holding target = holdingRepository.findHoldingByPortfolioIdAndStockCode(pfId, code).orElseThrow(HoldingNotFoundException::new);
 
         holdingRepository.delete(target);
@@ -163,7 +162,7 @@ public class PortfolioService {
         List<Holding> list = portfolio.getHoldings();
 
         // 보유 종목 없을시 바로 삭제, 아니면 보유종목 전부 삭제
-        if (list != null || !list.isEmpty()) {
+        if (list != null && !list.isEmpty()) {
             holdingRepository.deleteAll(list);
         }
